+++ c/src/pickle_printer.c
@@ -51,7 +51,11 @@ static void print_table_row(FILE* file, const PickleRow* pickle_row) {
 }
 
 static void print_pickle_table(FILE* file, const PickleTable* pickle_table) {
-    fprintf(file, "\"dataTable\":{\"rows\":[");
+    fprintf(file, "\"dataTable\":{");
+    if(pickle_table->argumentIndex > 0) {
+        fprintf(file, "\"argumentIndex\":%d,", pickle_table->argumentIndex);
+    }
+    fprintf(file, "\"rows\":[");
     int i;
     for (i = 0; i < pickle_table->rows->row_count; ++i) {
         if (i > 0) {
@@ -63,7 +67,11 @@ static void print_pickle_table(FILE* file, const PickleTable* pickle_table) {
 }
 
 static void print_pickle_string(FILE* file, const PickleString* pickle_string) {
-    fprintf(file, "\"docString\": {\"content\":\"");
+    fprintf(file, "\"docString\": {");
+    if(pickle_string->argumentIndex > 0) {
+        fprintf(file, "\"argumentIndex\":%d,", pickle_string->argumentIndex);
+    }
+    fprintf(file, "\"content\":\"");
     if (pickle_string->content) {
         PrintUtilities_print_json_string(file, pickle_string->content);
     }